Move required functions within spatialdata loader module
